55: extract `<Account>` and its test into separate files
