+++ frontend/src/App.test.tsx
@@ -15,7 +15,6 @@ import { createMemoryHistory } from "history";
 import { Router } from "react-router-dom";
 
 import { INITIAL_STATE, rootReducer } from "./store";
-import { Account } from "./App";
 
 // 2
 import { rest } from "msw";
@@ -41,37 +40,6 @@ import {
 
 // const BIG_VALUE_FOR_TIMEOUT_OF_ASYNCHRONOUS_OPERATIONS: number = 5 * 60 * 1000;
 // jest.setTimeout(BIG_VALUE_FOR_TIMEOUT_OF_ASYNCHRONOUS_OPERATIONS);
-
-describe("<Account>", () => {
-  test("renders the logged-in user's profile details", () => {
-    /* Arrange. */
-    const initState = {
-      ...INITIAL_STATE,
-      auth: {
-        ...INITIAL_STATE.auth,
-        loggedInUserProfile: {
-          id: 17,
-          username: "auth-jd",
-          email: "auth-john.doe@protonmail.com",
-        },
-      },
-    };
-    const realStore = createStore(rootReducer, initState);
-
-    /* Act. */
-    render(
-      <Provider store={realStore}>
-        <Account />
-      </Provider>
-    );
-
-    /* Assert. */
-    for (const text of ["17", "auth-jd", "auth-john.doe@protonmail.com"]) {
-      const tableCellElement = screen.getByText(text);
-      expect(tableCellElement).toBeInTheDocument();
-    }
-  });
-});
 
 /* Describe what requests should be mocked. */
 const requestHandlersToMock = [
